feat: refactor autotask names
